requirements.txt
Django==5.2.4
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==10.3.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pywebpush==1.14.0
cryptography==41.0.3
